Общее: 1. Запрос на выход из программы (задается в Настройках). 2. Change.rtf

+++ Source/Main/MainForm.cs
@@ -70,9 +70,20 @@ namespace Main
 		void TsbtnExitClick(object sender, EventArgs e)
 		{
 			// выход из программы
-			// очистка временной папки
-			filesWorker.RemoveDir( Settings.Settings.GetTempDir() );
-			this.Close();
+			if( ! Settings.Settings.ReadConfirmationForExit() ) {
+				// очистка временной папки
+				filesWorker.RemoveDir( Settings.Settings.GetTempDir() );
+				this.Close();
+			} else {
+				DialogResult result = MessageBox.Show( "Вы действительно хотите выйти из программы?", "SharpFBTools",
+			                                      MessageBoxButtons.YesNo, MessageBoxIcon.Question );
+	      	  if( result == DialogResult.Yes ) {
+					// очистка временной папки
+					filesWorker.RemoveDir( Settings.Settings.GetTempDir() );
+					this.Close();
+				}
+				return;
+			}
 		}
 		
 		void TsbtnFB2ValidatorClick(object sender, EventArgs e)
